requirements.txt
tkinter
cryptography
Pillow
opencv-python
requests
beautifulsoup4
ollama
sqlite3
